requirements.txt
# Dependências principais do projeto
python-telegram-bot
crewai
gspread
google-auth
google-api-python-client
python-dotenv
